Fixed a chest and key spawner
Spawns 3 chests with keys in and one key to the first key. No color variation on the chests but keys have names by color. Order of keys are always the same, so Yellow is last. But a nicer end hatch is needed that automatically detects last key so order can change. Also did som Interaction and HUD scripts just to test the system.